added sivenav for the view and a basic model for the layout
