+++ src/app/layout/layout.model.ts
@@ -1,0 +1,13 @@
+
+export interface AppView {
+    name: string,
+    i18n: string
+}
+
+export enum AppBarViews {
+    home = "home",
+    about = "about-us",
+    events = "events",
+    blog = "blog",
+    involved = "get-involved",
+}
