+++ src/app/layout/navbar/navbar.component.html
@@ -7,7 +7,7 @@
       </a>
       <ul class="left hide-on-med-and-down links">
         <li *ngFor="let view of views" [class.selected]="view === selectedView" (click)="onSelect(view)" >
-          <a routerLink="{{view.name}}">{{ view.translated | translate | uppercase }}</a>
+          <a routerLink="{{view.name}}">{{ view.i18n | translate | uppercase }}</a>
         </li>
       </ul>
       <div class="right">
@@ -18,7 +18,7 @@
 </nav>
 
 <ul appSidenav class="sidenav" id="mobile-menu">
-  <li>
-    <a class="sidenav-close">{{ 'links.about' | translate }}</a>
+  <li *ngFor="let view of views" [class.selected]="view === selectedView" (click)="onSelect(view)" >
+    <a class="sidenav-close" routerLink="{{view.name}}">{{ view.i18n | translate | uppercase }}</a>
   </li>
 </ul>
